+++ R/accuracy.R
@@ -122,11 +122,35 @@ percentile_score <- function(.dist, .actual, na.rm = TRUE, ...){
     mean(na.rm = na.rm)
 }
 
+#' @rdname distribution_accuracy_measures
+#' @export
+CRPS <- function(.dist, .actual, n_quantiles = 1000, na.rm = TRUE, ...){
+  if(is_dist_normal(.dist)){
+    mean <- map_dbl(.dist, `[[`, "mean")
+    sd <- map_dbl(.dist, `[[`, "sd")
+    z <- (.actual-mean)/sd
+    z <- sd*(z*(2*stats::pnorm(z)-1)+2*stats::dnorm(z)-1/sqrt(pi))
+    mean(z, na.rm = na.rm)
+  }
+  else{
+    probs <- seq(0, 1, length.out = n_quantiles + 2)[seq_len(n_quantiles) + 1]
+    percentiles <- quantile(.dist, probs)
+    if(length(percentiles[[1]]) > 1) abort("Percentile scores are not supported for multivariate distributions.")
+    z <- map2_dbl(percentiles, probs, function(percentile, prob){
+      L <- ifelse(.actual < percentile[[1]], (1-prob), prob)*abs(percentile[[1]]-.actual)
+      mean(L, na.rm = na.rm)
+    })
+    2 * mean(z, na.rm = na.rm)
+  }
+}
+
 #' Distribution accuracy measures
 #' 
 #' @inheritParams interval_accuracy_measures
+#' @param n_quantiles The number of quantiles to use in approximating CRPS when an exact solution is not available.
+#' 
 #' @export
-distribution_accuracy_measures <- list(percentile = percentile_score)
+distribution_accuracy_measures <- list(percentile = percentile_score, CRPS = CRPS)
 
 #' Evaluate accuracy of a forecast or model
 #' 
@@ -282,7 +306,7 @@ accuracy.fbl_ts <- function(object, data, measures = point_accuracy_measures, ..
   }
   
   grp <- c(syms(by), groups(object))
-  by <- union(expr_text(index(object)), by)
+  by <- union(index_var(object), by)
   
   
   if(!(".model" %in% by)){
@@ -290,13 +314,13 @@ accuracy.fbl_ts <- function(object, data, measures = point_accuracy_measures, ..
   }
   
   if(NROW(missing_test <- suppressWarnings(anti_join(object, data, by = intersect(colnames(data), by)))) > 0){
-    n_miss <- length(unique(missing_test[[as_string(index(missing_test))]]))
+    n_miss <- length(unique(missing_test[[index_var(missing_test)]]))
     warn(sprintf(
       "The future dataset is incomplete, incomplete out-of-sample data will be treated as missing. 
 %i %s %s", 
       n_miss, 
       ifelse(n_miss==1, "observation is missing at", "observations are missing between"),
-      paste(unique(range(missing_test[[expr_text(index(missing_test))]])), collapse = " and ")
+      paste(unique(range(missing_test[[index_var(missing_test)]])), collapse = " and ")
     ))
   }